DeepAR Framework Update

+++ ios/Classes/SwiftCameraDeepArPlugin.swift
@@ -257,11 +257,11 @@ public class DeepArCameraView : NSObject,FlutterPlatformView,DeepARDelegate{
                             if let parameter = (dict["parameter"] as? String){
                                 if let floatValue = (dict["floatValue"] as? Double){
                                     let f = Float(floatValue);
-                                   self.deepAR.changeParameter(changeParameter,component:component,parameter:parameter,floatValue: CGFloat(f));
-                                   NSLog("%f",f);
-                                   NSLog(changeParameter);
-                                   NSLog(component);
-                                   NSLog(parameter);
+                                    self.deepAR.changeParameter(changeParameter,component:component,parameter:parameter,floatValue: f);
+                                    NSLog("%f",f);
+                                    NSLog(changeParameter);
+                                    NSLog(component);
+                                    NSLog(parameter);
                                     
                                 }
                             }
